refactor: move secret drivers into plugin package
