+++ pkgs/standards/peagen/peagen/plugins/__init__.py
@@ -26,7 +26,7 @@ GROUPS = {
     "git_filters": ("peagen.plugins.git_filters", object),
     "vcs": ("peagen.plugins.vcs", object),
     "selectors": ("peagen.plugins.selectors", object),
-    "secrets": ("peagen.plugins.secrets_drivers", object),
+    "secrets": ("peagen.plugins.secret_drivers", object),
     # template sets remain in the top-level package
     "template_sets": ("peagen.template_sets", None),
 }
