+++ pkgs/standards/peagen/peagen/plugins/secret_drivers/__init__.py
@@ -1,0 +1,5 @@
+from .env_secret import EnvSecret
+from .base import SecretDriverBase
+from .autogpg_secretdriver import AutoGpgDriver
+
+__all__ = ["EnvSecret", "SecretDriverBase", "AutoGpgDriver"]
