--- pkgs/standards/peagen/peagen/plugins/secrets_drivers/__init__.py
@@ -1,3 +1,0 @@
-from .env_secret import EnvSecret
-
-__all__ = ["EnvSecret"]
