--- pkgs/standards/peagen/peagen/secrets/__init__.py
@@ -1,5 +1,0 @@
-"""Secret driver implementations."""
-
-from .driver import AutoGpgDriver, SecretDriverBase
-
-__all__ = ["SecretDriverBase", "AutoGpgDriver"]
